update mock_customer_es fixture
